refactor port stat setup to skip queue counters on disabled ports
Summary:
This refactors how we setup counters to always be part of the
program() function in bcmPort. Previously, we would directly call in
to the cos queue manager to create stats. Bypassing BcmPort makes it
difficult to disable the queue stats when we disable a port.

We still have one callsite where we will enable stats w/out having an
actual switch state node. I will get rid of this in a subsequent diff
when we add handling for ports added to the switch state.

Differential Revision: D16955606

fbshipit-source-id: 5d7dc75b8deec33e430635711f247e47c8ab6be0