have BcmPort::getPortStats return an optional
Summary:
Since we now can disable stats on a port when it gets
disabled, it is a bit disingenuous to return port stats when stat
collection is disabled. Before this we would keep returning the last
saved stats as they became more and more stale. This allows signalling
to users that the stats are not really valid anymore instead of making
calculations blindly based on zeroed out stats.

This is not ideal, but was a temporary measure to prevent BstStatsMgr
from exporting stats on ports without stat collection enabled and from
trying to export stats before stats were available.

Reviewed By: sas

Differential Revision: D16955610

fbshipit-source-id: 1f4a77494b1aa84f5f5ddb5c3be49d65ed36638b