add processAddedPorts to BcmSwitch
Summary:
Not having a processAddedPorts function was a source of
complexity during warm boot, especially once we started recovering the
entire switch state from the dumped file. This makes it so that we run
through the program() function even for the first state. This allows
us to get rid of some hacks for recovering mirrors and initializing
statistics.

This does open us up a bit to more risk in that we run through a
potentially dangerous function on warm boot. It does mean that if we
added some code that does something disruptive in to program(), we
could cause issues in warm boot. If we find this is too risky, we can
get rid of program and just enable the stats, though we will need to
add back in the mirroring hack.

Differential Revision: D16955607

fbshipit-source-id: 9ce5f463a4a4ed56f89ed25e108b65a69cbc869b