Deprecate storing portName in BcmPort
Summary:
Now that we store the entire port switch state node, there
isn't really a need to copy the name in. This changes all the call
sites that used to implicitly use portName to take the name as an
explicit argument.

Next phase (which I have talked about for a while) is to prevent port
name changes and just get a static port name from the platform. That
will help reduce some complexity to handle port name changes.

Reviewed By: sas

Differential Revision: D16955611

fbshipit-source-id: ddaf4803dff7b070e0f4c46c6373f9c14b3d7507