Emit cpu, rss usage for hw benchmarks
Summary:
As titled. For each benchmark, emit cpu time spent and max RSS
consumed in running that benchmark. I chose to measure this over the
entire benchmark run vs a fine grained operation like ECMP shrink.
The idea behind this is to catch a wider swathe of BCM switch and
SDK resource usage regressions in addition to the time based benchmarking
of specific operations (routes add, ecmp shrink etc).

As a illustration, this catches the CPU regression in 6.5.16 that
we have been battling recently. A run of ecmp shrink benchmark with
resource usage shows a ~30% increase in CPU time for SDK 6.5.16, even
while the ECMP shrink time remains unchanged P106440984

We would still need to carefully pick CPU, memory benchmark thresholds
so as to not get noise. That will be determined experimentally.
When we get a new SDK, the guideline would be to use the same thresholds
as before and then alert on any regressions.

This only covers the bottom half of our code, a similar approach should
be extended to top half benchmarking as well.

Reviewed By: shri-khare

Differential Revision: D17081976

fbshipit-source-id: e191980c27a6333cea3e8a992b90983f0e5939ad