Enable L2 mac address aging to 5mins
Summary:
See discussion here:

https://fb.workplace.com/groups/netstigations/permalink/2691669807548284/

but our L2 MAC address tables were found to be full in ~100 switches around
the fleet.  Why they were full is a different issue, but it exposed that
we did not have MAC address aging enabled.

This patch enables it and provides a BRCM test to make sure it works.

This is a very standard behavior in ethernet networks and there could
be problems (e.g., if there was a hardware swap and we couldn't learn the
new l2 addrs) if we don't enable it.

*But* it turns out this triggers a race condition in the SDK, see
T52790110 and CS8800124, so also turn off aging right before shutdown.

Reviewed By: boryas

Differential Revision: D16947258

fbshipit-source-id: 1640bdf4deddb39d1d0c25a2145e078cd8c770a4