Break down RouteDistirbutionGenerators into more modular pieces
Summary:
In the current version of RouteDistributionGenerator
we generate a vector of SwitchStates for a given prefix distribution.
This works well for h/w route programming benchmarks but would
not work well for rib/SwSwitch benchmarks where computing the
switch state entails doing all the work of route resolution (which
is what we wish to benchmark). So breakup into 2 route distribution
generators
1) A generator which generates a vector of vector of Routes for a
given distribution
2) A generator like before, which generates a vector of switch states.
2 is implemented via 1.

For now these are poorly named as RouteDistributionGenerator2. Subsequent
diffs will update to better names and update existing references.

Reviewed By: phshaikh

Differential Revision: D17001542

fbshipit-source-id: 50eacd5afbfe36cffa38a5c690ebca23051f44d5