Correct (logical) const correctness
Summary:
In RouteGenerators, allow member cache to be mutable,
since this does not represent the logical state of a generator.
Its essentially a optimization that we could do w/o. A downside
of not marking these mutable was that we had to make get api
non const, while per logical constness it should have been
const.

Minor other fixes as well.

Reviewed By: sas

Differential Revision: D17048816

fbshipit-source-id: 3d8777d067fde4bd2ae6d05df5e85c9df7328732