Redefine logic for BcmPortGroup::LaneMode enum
Summary:
Before, the LaneMode enum corresponded to the number of ports in a group. To get the lane mode, first the number of lanes needed per port was calculated, which mapped to the LaneMode needed (4 lanes: single mode, 2 lanes: dual mode, 1 lane: quad mode). However, this logic makes it hard to add and support more lanes (like for Wedge400DQ, which needs 8 lanes).

Thus, these changes redefine the meaning of the LaneMode enum to indicate the number of lanes needed per port, instead of how many ports are active in a group (4 lanes: quad mode, 2 lanes: dual mode, 1 lane: single mode). Then, when we support 8 lanes for Wedge400DQ this will just be 8 lanes for octal mode.

Differential Revision: D16847251

fbshipit-source-id: 2683d316530c3f7fb734b3808fabb4577d562c3f