allow different loopback modes across platforms
Summary:
Looks like PHY loopback mode does not work for TH3. This adds
a new property on BcmTestPlatform for the desired loopback mode. All
platforms outside of Wedge40 use MAC.

Kind of a pain to access the correct platform in the LinkToggler class
so this has a lot of threading stuff around.

Differential Revision: D17097112

fbshipit-source-id: d24fa6e2ef6be51db582ff0d636f4eeef6cc3250