add SaiApi support for sai_get_object_count
Summary:
During a warm boot, we will need to traverse the current hardware
state, regardless of exactly the higher level design. This is the first
diff towards that goal. saiobject.h defines a set of APIs keyed by the enum
sai_object_type_t which allow a user to do traversal fetches. The first, and
simplest is sai_get_object_count which simply gives us the number of objects
of that type. This diff adds all the support for the api in sai/fake and
implements it fully for PortApi, along with a test.

Differential Revision: D15816930

fbshipit-source-id: 1d4fbb9662cdf16b3cabf0e5c38f538409838a60