add SaiApi support for sai_get_object_keys
Summary:
During a warm boot, we will need to traverse the current hardware
state, regardless of exactly the higher level design.

After getting the count, the next step is to discover the available,
programmed "sai keys" by object. For most objects, this will be sai_object_id_t
but for some (route, neighbor, fdb) it will be an entry struct. Once we
recover the keys, we can start recovering the objects themselves

Differential Revision: D16386028

fbshipit-source-id: 942563795f30fef2f6054a1bed7b55e1a741387a