Introduce SaiObject for unified Warm Boot/Delta processing
Summary:
In order to generalize much of the code that is shared between delta
processing and warm boot recover/replay into SaiStore, this diff introduces a
generic "SAI Object RAII Type": `SaiObject`. The class is parameterized by a
SaiObjectTraits (e.g. SaiPortTraits, SaiNextHopGroupMemberTraits) and will
support:
* load existing by adapter key
* create new by adapter host key + attributes
* set new attributes by adapter host key + attributes

This type owns the lifetime of a SAI object: when SaiObject is destroyed, if
it currently manages a live SAI object, that object is removed from SAI. For
this reason, it cannot be copied, but it can be moved from. Most practical
users of this class will have a ref-counted shared_ptr so they don't have to
worry too much about moves/copies, that's more a consideration for SaiStore.

Differential Revision: D16818102

fbshipit-source-id: a4a75265d1bb4cb07c4b044a974c6e421cc4ec12