Extend LogThriftCall to accept log level
Summary:
We have some thrift calls that are constantly spammed and I
want to skip logging for them. However, the current implementation
hardcodes using XLOG(INFO) internally. This refactors the code and
adds a macro so that we can pass in log level similar to XLOG.

I also fixed the line num and file name parameters to come from the
callsite, no longer from inside the LogThriftCall.cpp file.

Differential Revision: D17138624

fbshipit-source-id: 14d8fb77a8fb8d6f89c2dfae56d3223bdf22e8c9