Move getPreAsicInitFile API out of platform and move to BcmSwitch
Summary:
We don't need to customize this file path based on each
platform. We may need to customize the file we pass to each
platform, but that is easily controlled via contents of this
file. If we ever needed to distinguish pre ASIC init  scripts
for different platforms we could easily do so by naming these
files differently for different platforms.
At any rate we have only ever needed to pass in a pre init file.
Having it just be defined and used inside BcmSwitch suffices.

Reviewed By: xzhu70

Differential Revision: D17177124

fbshipit-source-id: 483cf32cf0f0eaa405b253eaf691dff78474b158