implement SaiVlanManager in terms of SaiStore
Summary:
add a "set" API to SaiStore which creates a new object
or overrides it with new attributes.

Use this new API to implement SaiVlanManager on top of SaiStore.
Of particular interest is the SaiVlanHandle and the removal of
SaiVlan, SaiVlanMember. The latter two are replaced with the totally
generic SaiObject<> from the store/ library. The creation logic stays
largely the same, and unfortunately still depends on other managers because
we can't lookup a Port by any other means currently. I have some ideas for
reconsidering that assumption, though.

Differential Revision: D16701666

fbshipit-source-id: e1a37adb861f0aa285a4a49b7cd60d08e4ac2402