Migrate flex port modes to SwitchEnsemble from BcmTest
Summary:
Again, like logical port ids, this is a platform/ensemble
property and not a Test property
While at it also change list<FlexPortMode> to vector<FlexPortMode>
to be consistent with other uses in tests (and get better locality
of reference as a upshot). There is no compelling reason why
this should have been a list in the first place.

Differential Revision: D17101231

fbshipit-source-id: 2e4cb50c261e1f2bfc0625ee623917c94104e6ac