skip interface link local routes
Summary:
Same link locak ip address is used on all fboss interfaces. When a RIF is created, IP2TOME routes will be added. The same link local route is added for every interface which is causing a duplicate route add.
This diff skips adding the route.

The link local interface route will be added as part of the ACL.

Differential Revision: D17230804

fbshipit-source-id: af11c13b49988a728c512ff00f00f5bda2435ba1