Move thrift thread setup to Hw/BcmSwitch ensemble from Hw/BcmTest
Summary:
Mostly a code move, to move responsibility of setting
up thrift handler to Hw/BcmSwitchEnsemble from Hw/BcmTest. The
idea here is to move nearly all of setup responsibilities to
Hw/BcmSwitchEnsemble and not in test setup. Motivations for this
are
- Hw/BcmSwitchEnsemble get used in contexts outside of tests. So
for example after the above change we would have shell access to
debug our benchmark code as well.
- Sai tests - Hollowing out BcmTest class means that more tests
can derive from just HwTest class and be shared across SAI and BCM
platforms.

Reviewed By: phshaikh

Differential Revision: D17119952

fbshipit-source-id: 298e5a18bda8f53ad401511df1d2698bbf29bd72