Implement SaiNextHopGroupManager in terms of SaiStore
Summary:
This brings us support for mixed normal/warm-boot operations on next
hop groups.

N.B., this is actually sort of awful. My high level idea was that
SaiStore being a RefMap (works well for WB + clear cache case) would also
automagically be a great fit for next hop groups. Unfortunately, it turns out
we need a RefMap of SaiNextHopGroupHandle since it is the group AND members that
must be referenced, and modified in place "underneath" routes.

I have an idea for how to add functionality to SaiStore to support this nicely,
(pub-sub callbacks and/or coroutines to implement nbr resolution makes the
lifetimes a lot cleaner, I think) but in the interest of making progress, this
diff is a compromise.

Differential Revision: D16949063

fbshipit-source-id: 2510a336a5dea1bf34589da65a19c89fe229dfca