Move responsibility of creating HW ensemble to a standalone function
Summary:
The idea here is to provide a hook for various ensemble
implementations (SAI, BRCM) to provide a implementation to create
a SwitchEnsemble at runtime. Previously we were doing this in
the derived BcmTest class. Doing so, made it that all tests needed
to derive from BcmTest class, even if the could work using just
the HwEnsemble/Switch interface. This diff attempts to fix this
limitation. In the new scheme of things
- Implementations (SAI, BRCM) would provide their own creatHwEnsemble
implementation. You can think of SAI providing a implementation which
creates a SaiSwitchEnsemble (SaiSwitch, SaiPlatform) and a bcm implemetation
will create a BcmSwitchEnsemble.
- For the immediate use case of sharing test code, this paves the way
for sharing all tests which interact with Hw via the HwSwitch/HwSwitchEnsemble
interface. This should cover all tests that don't care to look at the
internals of BcmSwitch layer data structures. To achieve the full potential
of this we will need to build further generic utilities - say for e.g. one
to get the width of a multipath group for a route. Those would again have
specific implementations for BCM/SAI layers, but can provide a uniform
interface.
- Moving further, I see we would have 3 classes of tests
 - Generic HW switch tests
 - BCM specific tests
 - SAI specific tests

Its not yet clear where each of the existing tests will fall. But
this provides us a way to start investigating

Reviewed By: phshaikh

Differential Revision: D17124972

fbshipit-source-id: 46274581cc672fe1984eb81477caf45ea7259540