Port BcmVlanTest to be hw agnostic HwVlanTest
Summary:
Building on the work done in diffs earlier in
the stack. Abstract out vlan tests that are implemented
in terms of generic vlan querying APIs viz.
- std::vector<VlanID> getConfiguredVlans(const HwSwitch*);
- std::map<VlanID, uint32_t> getVlan2NumPort(const HwSwitch*);

As long as a Hw implementation (BRCM, SAI) plugs in this implementation
we can run these vlan tests on it.
Other things to note
- I got rid of the ports querying in these tests. These functionality
that ports get enabled in h/w after we apply such a switch state is
tested indirectly in many places. Plus if we ever wanted to test this,
Port tests are the more appropriate place
- I had to set link_whole for the hw_switch_test lib. Without this,
the HwVlanTest file would get dropped since nothing in the application
directly refers to it. This wasn't a problem when all tests were part
of bcm_test binaries

Differential Revision: D17142601

fbshipit-source-id: 58bd317b3adc1c5f691bd1f2ae5f9bfc00d97cdc