populate trunks from warmboot only if trunks key exist
Summary:
We are looking for "trunks" key in the warmboot cache which does not exist, causing the agent to core (P110286774). This is becuase we are trying to warmboot from an agent which does not have support for trunk warmboot to agent which has the support.

During warmboot, make sure if the "trunks" key exist in the cache before accessing it.

We can delete this once the agent is rolled out with this code change.

Reviewed By: aeckert

Differential Revision: D17340746

fbshipit-source-id: b4b521498c35c8df6b18ae4a85fc947ee6b4404d