mirror manager with source ip
Summary:
mirror manager while resolving tunnel attributes to use source address setting if provided in configuration.

once resolved tunnel has source address settings done, it will be used in underlying BCM switch to setup mirror with this source address.

if source ip setting is not set, an address present on interface which is used to send mirrored traffic is used.

Differential Revision: D17356392

fbshipit-source-id: 51751f5397d89df4ea7f549056915521c55bc994