Add HwLinkStateDependentTest on the lines of analogous class in Bcm layer
Summary:
Add HwLinkStateDependentTest on the lines of analogous class in Bcm layer.
Tests which depend on link states - ecmp shrink, all dataplane tests,
which need to send packets over UP links to derive from such a class.

Initially I was thinking of merging this functionality into HwTest
itself. I didn't go down that route since
- LinkStateDependent tests impose interface requirements - in particular
a starting config must be provided.
- LinkStateDependent tests starting state is application of aforementioned
config and getting all ports up (to synchronize all port events that may
occur as result of initing the chip). This may not be desirable for all
tests. For e.g. we have tests that assert the chip default state post
a cold boot. One of the assertions there is that all ports are disabled
on a cold boot.

Reviewed By: sas

Differential Revision: D17331719

fbshipit-source-id: ae3795d3396bd4c01c37e8958c001f9d030429fd