Our first HwSwitch agnostic dataplane test
Summary:
This diff port trunk load balancer tests to
be HwSwitch agnostic. The main trick is to satisfy the
functionality through the HwSwitchEnsemble and other
abstractions worked out earlier.

We don't have hashing or LAG functionality yet in SAI.
But when we do, the test will be there to help us validate.

Reviewed By: phshaikh

Differential Revision: D17334924

fbshipit-source-id: b29b6ec28845f291b9814cfecf315d02394b2be6