Add Sai link state toggler
Summary:
As titled. Along the lines of BcmLinkStateToggler. There
is one key piece of functionality lacking here - which is to set
preemphasis value for ports. I am not sure how to do this for SAI
yet. In any case, the preemphasis value setting is only important
when we are bringing ports down, port up should still work. Will
figure something out for bringing ports down.

Reviewed By: phshaikh

Differential Revision: D17406264

fbshipit-source-id: 154e845f129a4f25a0986d09941bd099d6b0352c