fetchL2Table implementation for SaiSwitch
Summary:
Fill in fetchL2Table implementation for SaiSwitch. This API
is used via thrift and in mac learning tests to get the l2 table for
switch. Implementation here is pretty simple - getObjectKeys for FDB
entries and use those to fill in L2 entry thrift.
Caveats
- Handling LAG/Trunk ports is a TODO for when we implement LAG ports
- This does not yet work, since SAI implementations we use don't have
getObjectKeys implemented for FDB. But getting the code in while I am
at it. Mac learning tests will be used to verify this once the necessary
SAI SDK functionality is in

Reviewed By: phshaikh

Differential Revision: D17430520

fbshipit-source-id: 21e18ae30d0f43c6591c347d29608852174e4da1