Add a proxy class for NeighborUpdater
Summary:
This will be used to move all neighbor updater operations to a single thread.
For now, it's just a simple method forwarder that doesn't do anything special.
I put it in a separate diff to make the review easier.

The `stateUpdate()` method is also purposefully left in the `NeighborUpdater`
class because state update processing will *not* be done in the neighbor
updater thread. Instead, state updates will be processed in the update thread,
and the resulting changes to neighbor caches will then be queued on the
neighbor thread and applied there. An example of this model can be seen with
`timeoutsChanged()`, which is a forwarded method and is called from
`stateUpdated()` after we determined that some changes need to be applied.

Reviewed By: aeckert

Differential Revision: D16744002

fbshipit-source-id: dd70da66a65dcb28d80f071febe267439f5eb000