support async packet tx in SaiSwitch
Summary:
SAI callback based tx/rx has no provision for asynchronous packet
tx. As a result, we need to implement it ourselves on top of the synchronous
packet tx function. We opt for the implementation that most closely matches
similar situations in FBOSS: a dedicated thread managed by an event base that
we dispatch the work to.

Differential Revision: D15700587

fbshipit-source-id: cba5f3d5d8f3e2003489f918fd42b2a646979c29