sai rx callback bottom half
Summary:
FBOSS rx depends on getting a PortID and a VlanID from the rx
packet. Currently, BcmSwitch can rely on the raw value from the callback,
but due to the indirection inherent to sai ids, we cannot. As a result,
we need to do a lookup of PortID and VlanID by sai_object_id_t. This races
with changes to those structures so it needs a lock. Taking that lock can
block, stalling the underlying rx thread and building up hardware buffers
for no reason. For this reason, we put the work that needs to look at
SaiSwitch internals onto a separate bottom half thread, as inspired by
the link scan bottom half thread in BcmSwitch.

Differential Revision: D15700586

fbshipit-source-id: 8c582f26ea0f93e8231cda48bf2bca70a41917f6