Consume new PortQueueRate
Summary:
For the first agent push with this patch, a warmboot will not reinstate packetsPerSec,
but the config will subsequently reinstate pps to desired value using the new
PortQueueRate fields. The time window is very small, and thus we don't handle it
explicitly in the code. Subsequent agent pushes will not suffer from this behavior as
PortQueueRate is dumped to switch state and will be reinstated from there.

Differential Revision: D17303859

fbshipit-source-id: f5b8fe2c7270e68a8efc94d7fc552b141f9c26a5