Fix vlan deletion race condition in NeighborUpdater
Summary:
This is old code that used to make sense when the threading model of the
neighbor updater wasn't as well defined. Now that we run everything in on a
single thread we don't need to take these extra precautions that I'm deleting
in this diff. More than that, these extra precautions were causing a deadlock
on exit, which is now fixed.

Reviewed By: aeckert

Differential Revision: D17588360

fbshipit-source-id: 58de110719dfc72290625a3eb0ebde9cf3da5249