Change kProfileValues to be a map of str->str rather than char*->char*
Summary:
There was a bug in how we were passing the configuration file
location as a profile value. We used the return value of
SaiPlatform::getHwConfigFileLocation (which returns a std::string)
and extracted a c_str out of it. This has obvious lifetime problems and
we paid for this in BRCM sai init which uses it. Fix this by changing
the map type

Differential Revision: D17677140

fbshipit-source-id: 12e5336198ab67422df00a457a05a89902cb9e81