Apply config for Rx Reason to queue
Summary:
This diff enables RxReasonToQueue to be propagated from the config to the hw switch.

Currently, we do not act upon this in BcmSwitch. I am enabling this for SaiSwitch so that these can be driven from config instead of programming them directly in the HwSwitch. This creates the Hostif rules to program reason code to queue mapping.

After SAI copp, I can fix the bcm also to use the config.

Few concerns: This is not a clean way to enable it since RxReasonToQueue is optional.

-  CPUtrafficPolicy is optional and also RxReasonToQueue is optional. Should we mandate these fields and generate the config all the time ?
- Other approach is to move RxReasonToQueue out of CPUtrafficPolicy which would involve config change.

Differential Revision: D17609987

fbshipit-source-id: 83c61da03395fc2eb75c96e025435153399dbcf7