s/multiplePortInVlanConfig/oneL3IntfNPortConfig
Summary:
Similar to D17715796, multiplePortInVlanConfig was
not mirroring production configs correctly. In that
i) It was putting ports in default VLAN
ii) It was not creating corresponding L3 intf for vlan (we would
get by since this is allowed for default VLAN but not for any other VLANs).

oneL3IntfNPortConfig does both these correctly and also fulfills the intent
of what we want done here.

Reviewed By: sas

Differential Revision: D17717178

fbshipit-source-id: 6ff9fc94c49842c5674a86cda276aefcaa50004d