Blow away interface addresses as well before destroying HwSwitch
Summary:
BRCMs ALPM mode requires that default routes be added first and
destroyed last. So to fulfill that requirement, before destroying
HwSwitch, we emptied the route tables of everything except default
routes. This worked fine on BRCM native SDK implementations.
On SAI, routes come from switch state route table as well as
from interface IPs (via interfaceToMe routes). So to satisfy
this requirement, blow away interface addresses as well.

Note that this is done just prior to deleting the HwSwitch,
the above work is just to satisfy ordering requirements. Its
not the best design since BRCMs ALPM requirements move up to
HwSwitchEnsemble  - but this shortcut was taken with the idea
that we only delete HwSwitch in tests, so rather than refactoring
a whole lot of BcmSwitch and SaiSwitch code (SaiSwitch would be
even harder since routes come from 2 sources), just do the
cleanup required for ordering in tests

Reviewed By: sas

Differential Revision: D17705048

fbshipit-source-id: 36588ceb4a5495726a06a1f4074f0c402ecf1e46