Update default vlan to 4095 in hw test configs
Summary:
Primary motivation here is to work around
lack of get_object_count/keys support in brcm-sai right
now. brcm-sai creates vlan 1 and puts all ports in it
on cold boot. Then when we apply the config, we try
to add VLAN 1 again. This will be fixed once we/brcm
implements support for the 2 APIs above.
This change should be benign for bcm_tests and as a
+ve side effect we get better mirroring of production
configs - there we pick 4095 to be the default VLAN
and not 1.

Reviewed By: sas

Differential Revision: D17708742

fbshipit-source-id: 079b568ede23b6f190f77a2b4f868f468e2db724