first take on limited concurrent access to data in SaiSwitch
Summary:
packet rx needs to lookup port and vlan ids. This needs to be
done concurrently with arbitrary (and arbitrarily slow) updates of the
SaiSwitch. To facilitate this, we use a ConcurrentHashMap for only this
data. If the iterator in the rx thread is out of date, there is no real
harm: we might drop a packet. This same assumption exists elsewhere in
the rx pipeline.

One open question is whether this centralized design for ConcurrentIndices
helps or hurts. I think it explicitly calls out what is happening, but it is
admittedly a bit clunky. Very open to feedback on this design, naturally.

Differential Revision: D15750266

fbshipit-source-id: 96a71b3024bf817302c8e5f314de7b4a65eaa6cf