Unregister all callbacks not just rx
Summary:
In unregisterCallbackLocked, unrgister - rx, port and FDB event
callbacks, not only RX.
Secondly, for terminating the rx thread loop, do eventBase.terminateLoopSoon
and join w/o holding the lock. See discussion on D17695705 for a example.
Post concurrent indices diff we would not hold a lock in rx callback
path, but its still good hygeine to not bank on that.

Reviewed By: boryas

Differential Revision: D17772327

fbshipit-source-id: e71de21fdbd10e495faaef75f833b5a9a47766a0