stop accepting thrift connections before shutting down
Summary:
Saw one crash on a getCounters call during shutdown. This
looked to be deep in the guts of fb303. I will do some due diligence
debugging that, but it seems like we can avoid this class of issues
in general by stopping accepting all thrift connections when we
receive a signal.

Reviewed By: sas

Differential Revision: D17730985

fbshipit-source-id: f61cc8473ad7698da282d454f11f8b5e0001580f