Get rid of "fake warmboot" testing
Summary:
This was used to test warmboot when running tests against the fake SDK. It
mostly works, but its behavior started diverging from real warmboot behavior so
we cannot keep it in its current state. The way forward would be to either make
the behaviors of these two paths consistent, or to get rid of warmboot testing
with fake SDK completely. This diff goes with the former approach.

Differential Revision: D17858564

fbshipit-source-id: 373ea734671f86c3fa78aa4854ffb652fdcbacd5