re-organize low frequency port recovery ops
Summary:
We have two operations that we do if a port has been down for
a while--ensuring tx is enabled and then flapping the power class.

This was done in a somewhat ad-hoc way before and this organizes them
in to a new "remediate" function. It also makes sure that we only ever
do this if it has been more than 5 minutes since the last time the
port was up. This will avoid taking either of these actions if the
port has only been down for a short period of time.

Reviewed By: aeckert

Differential Revision: D14453228

fbshipit-source-id: a35c93d97b877c6a08c6b2cf73d91e864a2e5f17