Add a Bcm Tx slow path benchmark
Summary:
This is the flip side of rx slow path benchmark
added in D17785294. Methodology here is
- Program a route
- Start a thread and have it pump traffic till its stopped
- Get port out bytes, pkts.
- Burn 5 secs
- Get port out bytes, pkts.
- Calculate rate.

Interestingly the rate on TX is only 15K pps compared to 100K+
on the RX side. This is what routing folks were seeing when they
measured the slow path performance and got a cap at ~11K.

Reviewed By: saifhhasan

Differential Revision: D17839621

fbshipit-source-id: bc2e86301f1f445c0aa81bb866f410b49fa10ba1