Tell LSAN of deliberately leaked objects on warm boot exit
Summary:
During warm boots we deliberately leak the HwSwitchEnsemble
so as to not run all destructors (which would cause entries to
be removed from tables). Tell LSAN to ignore this for leak
calculations.
There is a tradeoff here though - in that SDK may also leak
objects deliberately during the warm boot path and we will need
to account for those. If this becomes onerous, we can always
fallback to using _exit on warm boot path.

Reviewed By: sas

Differential Revision: D17925799

fbshipit-source-id: 2b58ccff1a7678e12e031aa0e6f18a9c394a418b