Fix memory leak in BcmSflowExporter
Summary:
When we couldn;t get a address from fbwhoami we fallback
to getting one from the local interfaces. Here we use getifaddrs
which returns a dynamically allocated list of ifaddr structs.
This must be freed when done
http://man7.org/linux/man-pages/man3/getifaddrs.3.html

Detected via LSAN

TODO: Look in common or folly to see if something like
this already exists and use that rather than handrolling
our own.

Reviewed By: sas

Differential Revision: D18024503

fbshipit-source-id: c22d93fb34e97eddc7299363f753fb917c86fd22