Enforce the use of `override` for methods
Summary:
When a child class overrides a virtual method of a parent class, it is possible
to ommit both the `virtual` and the `override` keywords, which will still
produce an overriding virtual method, but with no visual indication of that to
the reader. Enforcing the use of `override` makes the code more readable.

Reviewed By: boryas

Differential Revision: D18071511

fbshipit-source-id: 8341bbace4f42b743136cfbd0db764fe8334410e