State observer to configure Queue-per-host for host entries
Summary:
This diff implements a state observer viz. LookupClassUpdater.
When a host entry is resolved, agent knows the MAC address and the
associated port. The state observer associates a classID with that host entry.

If multiple host entries resolve to same MAC, the same classID is associated
with each of those host entries. This could happen when connected server has
IPv4 (ARP resolution) as well as IPv6 address (NDP resolution).

Subsequent patches add HwSwitch logic to consume this classID from the
SwitchState and program the ASIC.

Differential Revision: D17254465

fbshipit-source-id: 8393ff736c0412e515bf235cb01c0e95a5cd31f7