Get queue gport based on port
Summary:
Currently, we fetch all the gports of queues of ports and cpu at once in BcmSwitch::setupCos() during initializing. This works on the case that we don't have to create or remove a port after initializing. But for Tomahawk3, we can support add/remove ports with the new PortResource apis. Therefore, we need to support getting the queue gports back after adding a new port.
To do so, I move the whole traverse call back logic back to BcmCosQueueManager initialization, which will be called during BcmPort initialization.

Reviewed By: aeckert

Differential Revision: D18072588

fbshipit-source-id: 659813244d2c5009c67bd8296183e6c23a962f98