Standalone RIB route programming benchmarks
Summary:
This benchmark compares the old route update mechanism to the new one.

Route generation is not excluded from the benchmarked code because the route
generation itself is presumably the same between both benchmarks, but in the
case of the old route update mechanism, the route generator also creates the
intermediate switch states, which is really what takes time when updating
routes, and can't be skipped.

Differential Revision: D16069872

fbshipit-source-id: f427889eb70047671cc1550e477518e0c1045748