Merge ClientID and StdClientId enums
Summary:
It seems like the split comes from the fact that thrift enums were not strongly
typed (no `enum class`) in the past, and we used `FBOSS_STRONG_TYPE` to define
`ClientID` and get some type safety. This is not the case anymore today,
`StdClientId` was already a strongly typed enum so we can just use that type
directly instead of having conversion functions between both.

I also renamed `StdClientId` to `ClientID` to keep the old internal API and
have cleaner/shorter names.

Reviewed By: boryas

Differential Revision: D18023748

fbshipit-source-id: 69fd5193d5df7007a52bf149f1ce50d7c0f8f9a9