Uniform specification of HwSwitch feature flags
Summary:
`FLAG_A & ~FLAG_B` is equivalent to just `FLAG_A`. Most places referring to
`HwSwitch` feature flags simply specify which features they need, so it'll be
nicer to make these uniform. Also, if we end up adding another `HwSwitch`
feature flag in the future, we'd end up with only part of the flags listed in
this exclusion pattern.

Differential Revision: D18204101

fbshipit-source-id: d006aa61b4871682777e4d7fbe2c389e067069e9