Replace dscp=48->hipri with stricter ACLs
Summary:
To prioritize NC traffic, I had added a acl to
catch dscp=48 traffic with a action to send packets to CPU
hi-pri queue. My (mistaken) assumption was that since this
was directing packet to CPU queue, ACL will only apply
to packets that trap to CPU. However the ACL applies always,
for front panel packets, it simply skips over the action since
the action directs packets to a CPU queue and the packet is
destined to a front panel queue (bcm_test to explicitly
cover this was added). Problem happens when we add ACL for
actions on front panel port packets. There are 2 examples
of this
- ttld ACLs
- Newly added MH nic ACLs
In such a case, dscp 48 packets going out of front panel packets
will be matched by the earlier network control ACL which will then
shadow the later ACL, causing them to be skipped. Tighten ACLs
to be -
- dscp 48 + DstIpLocalV{4, 6}
- dscp 48 + Ipv6 link local mcast
- dscp 48 + Ipv6 link local ucast

This diff adds bcm_tests for the new policy, configerator diffs
coming soon.

Reviewed By: shri-khare

Differential Revision: D18289216

fbshipit-source-id: 384fe721c17638b552693be8cd667619f720f6af