Hw warm boot exit time benchmark
Summary:
Start benchmarking the time we take in HwSwitch
during warm boot exit. While this is not a perfect match for
how long it takes its close. It includes
- Time it take to dump switch state, sdk state
- Time it takes to stop hw switch threads, handlers, get ASIC ready
for shutdown
It does not include
- SwSwitch thread stop time
- Any additional SwSwitch cleanup

We should attack the missing pieces but the current diff gets us quite
a ways there in benchmarking our warm boot exit path.

Methodology
- For all but wedge40 platforms, program the worst case FSW route scale
distribution and initiate warm boot exit.
- For wedge40, program distribution taken from a random wedge40 in fleet
and initiate warm boot exit.

Reviewed By: boryas

Differential Revision: D18231531

fbshipit-source-id: 4d3d5fc43a58cf6e76c13bbd3a0f905ba742a43c