splitting config application and port bringup
Summary:
Currently, the ports are brought up during the INITIALIZED state and waiting on a lock for port up notification.
In Sai switch, the link scan cb is enabled after the CONFIGURED state and we wait indefintely on the lock. Splitting the config application and port bring up so that ports are brought up after transitioning into CONFIGURED state.

Differential Revision: D18321473

fbshipit-source-id: dde14956145e1772314df977db6c6ddf223a3969