De-duplicate FIB routes prior to programming
Summary:
After a dynamic route update, the previous implementation of
`ForwardingInformationBaseUpdater::createUpdatedFib` created a new FIB route
for each resolved RIB route. Because `NodeMapDelta` works on the basis of Node
identity (as opposed to Node equality), each route update would cause
`BcmSwitch` to examine _every_ route in the FIB- even those which did not
change across state updates.

To remedy this, `ForwardingInformationBaseUpdater::createUpdatedFib` is
re-written so that if a RIB route is unchanged after recursive resolution, the
FIB is updated with a _pointer_ to the FIB route corresponding to the RIB route
prior to the resolution.

Differential Revision: D15523685

fbshipit-source-id: 112e6bbb7dc2c6a1c8180984a15bc5899e63693e