Make sure enabling port after speed change
Summary:
D18354097 adding back the old logic only if the PortState changes from disabled to enable, we'll call enable. But it forgot to consider the case that portState is still enable but speed change from 100G -> 4X25G. In such case the controlling port speed changes from 100G to 25G, but it won't call the enable function.

In prod, we don't broadly change port speed from warmboot, that's why the current push:339 and fboss-dev oncall didn't receive any related task like this.

Therefore, to fix this, we enable a port as long as:
1) port state change from disabled to enable
2) speed change

Differential Revision: D18417975

fbshipit-source-id: dde7dd3a154eded4d8ae2f6c684c9d9732ed9a8e