Move Platform::initPorts into Platform::init
Summary:
This diff will be helpful to the new platform port config refactor.
- With the new platform config design, we don't have to hard-code all those logical ports and transceiver id mapping in the XXXPlatform. Instead we can use the new platform config to create the platform port mappings like D17814477.
- Besides, the new platform config will contain more information, such as PinConnection, lane_mapping, pim_id, transceiver_id, port name and etc.

Right now, for BcmPlatform, we only call initPorts() right before we start initializing BcmPortTable. If we want to use any new information mentioned above, we have to wait for BcmPortTable::initPorts() to be called.

Therefore, I decided to move Platform::initPorts() in Platform::init() to make initPorts as part of Platform::init() step.

Differential Revision: D18393572

fbshipit-source-id: 9e917f37dac0af2ca6c957540ffee0c9c231d3e8